libathemecore/function.c: random_string: use size_t not int for length arguments
This breaks the ABI on 64-bit but we're in the middle of a development
cycle anyway; master should not be considered stable.

+++ include/tools.h
@@ -129,7 +129,7 @@ typedef struct {
 } email_canonicalizer_item_t;
 
 /* misc string stuff */
-E char *random_string(int sz);
+E char *random_string(size_t sz);
 E void create_challenge(sourceinfo_t *si, const char *name, int v, char *dest);
 E void tb2sp(char *line);
 E char *replace(char *s, int size, const char *old, const char *new);
